+++ src/Manager/Manager.ts
@@ -3,9 +3,11 @@ import { Contract } from '../Contract/Contract';
 import { resolve, parse } from 'path';
 import { giveGrams } from '../Deploy/Deploy';
 import { Wallet } from '../Contract/Wallet';
+import { Helpers } from './Helpers';
 export default class Manager {
   public client: any;
   public contracts = {};
+  public helpers = new Helpers();
   // eslint-disable-next-line @typescript-eslint/no-empty-function
   constructor() {}
   public async createClient(servers: string[] = ['net.ton.dev']) {