+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file.
 - Now the loadContract function is `async` and calculates the future address of the contract with an entry in the `.address` field
 - Manager no longer generates and stores keys inside itself
 - loadContract signature
+- Manager helpers
+  -- deployCheck
 
 ### Remove